requirements.txt
requests
datetime
asyncio
discord.py